Separated ModelDescriptor and ModelAttributes.
