+++ h2o-core/src/main/java/hex/Model.java
@@ -5,8 +5,7 @@ import hex.genmodel.algos.glrm.GlrmMojoModel;
 import hex.genmodel.algos.tree.SharedTreeGraph;
 import hex.genmodel.algos.tree.SharedTreeMojoModel;
 import hex.genmodel.algos.tree.SharedTreeNode;
-import hex.genmodel.descriptor.Table;
-import hex.genmodel.descriptor.VariableImportances;
+import hex.genmodel.descriptor.ModelDescriptor;
 import hex.genmodel.easy.EasyPredictModelWrapper;
 import hex.genmodel.easy.RowData;
 import hex.genmodel.easy.exception.PredictException;
@@ -2642,10 +2641,6 @@ public abstract class Model<M extends Model<M,P,O>, P extends Model.Parameters, 
       public String uuid() { return String.valueOf(Model.this.checksum()); }
       @Override
       public String timestamp() { return new DateTime().toString(); }
-      @Override
-      public VariableImportances variableImportances() { return null; }
-      @Override
-      public Table modelSummary() { return null; }
     };
   }
