+++ h2o-genmodel/src/main/java/hex/genmodel/AbstractMojoWriter.java
@@ -1,5 +1,6 @@
 package hex.genmodel;
 
+import hex.genmodel.descriptor.ModelDescriptor;
 import hex.genmodel.utils.StringEscapeUtils;
 
 import java.io.IOException;
