+++ h2o-genmodel/src/main/java/hex/genmodel/attributes/VariableImportances.java
@@ -1,4 +1,4 @@
-package hex.genmodel.descriptor;
+package hex.genmodel.attributes;
 
 
 /**
@@ -7,6 +7,7 @@ package hex.genmodel.descriptor;
  */
 public class VariableImportances {
 
+    // Index is the shared key to both. A record under index {i} in variables is the name of the variable 
     public final String[] _variables;
     public final double[] _importances;
